requirements.txt
numpy==1.15.3
Pillow==5.3.0
PyOpenGL==3.1.0
